chore: refactor queue & profile code
