+++ pages/[slug]/queue.tsx
@@ -1,100 +1,16 @@
 import Image from 'next/image';
 import bokeh3 from '../../public/pfm-bokeh-3.jpg';
-import { useEffect, useState } from 'react';
 import NavBar from '../../components/Utils/NavBar';
 import QueueSong from '../../components/QueueSong';
 import LoadingSpinner from '../../components/Utils/LoadingSpinner';
-import { User } from '../../models/User';
-import { Song } from '../../models/Song';
 import Layout from '../../components/Utils/Layout';
-import { usePathname } from 'next/navigation';
-import { Bid } from '../../models/Bid';
-import { Play } from '../../models/Play';
 import { usePusher } from '../../components/hooks/usePusher';
-import { getUserProfileFromLocal } from '../../utils/profile';
-
-// pleb.fm/bantam/queue
-// Used for frontend hydration
-export type SongObject = {
-  trackTitle: string;
-  artistName: string;
-  feeRate: number;
-  playing: boolean;
-  myPick?: boolean;
-  upNext: boolean;
-  bidders: User[];
-  queued?: boolean;
-  status?: string;
-};
-export const fetchSong = async (
-  songId: string,
-  shortName: string,
-): Promise<Song> => {
-  const queryString = new URLSearchParams({
-    id: songId,
-    shortName: shortName,
-  });
-  const res = await fetch(`/api/spotify/getSong?${queryString}`, {
-    headers: { 'Access-Control-Allow-Origin': '*' },
-  });
-  if (!res.ok) throw new Error('Failed to search song');
-  const result = await res.json();
-  return result;
-};
-export const cleanSong = (song: Play, user: User): SongObject => {
-  const bidders = song.bids.map((x: Bid) => x.user);
-  const totalBid = song.runningTotal;
-  const myPick = bidders.some((x: User) => x.userId === user.userId);
-  return {
-    trackTitle: song.songName,
-    artistName: song.songArtist,
-    feeRate: totalBid,
-    playing: song.status === 'playing',
-    myPick,
-    upNext: song.status === 'next',
-    bidders,
-    queued: song.status === 'queued',
-    status: song.status,
-  };
-};
-
-export const getQueue = async (
-  host: string,
-  user?: User,
-  isProfile: boolean = false,
-): Promise<SongObject[]> => {
-  let url = `/api/leaderboard/queue?playing=${true}&shortName=${host}`;
-  if (isProfile) {
-    url += `&userId=${user?.userId}`;
-  }
-  const response = await fetch(url);
-  const res = await response.json();
-  if (!res?.data) {
-    return [];
-  }
-  const songs = res.data.map(cleanSong);
-  return songs;
-};
+import { useQueue } from '../../components/hooks/useQueue';
 
 export default function Queue() {
-  const [queueData, setQueueData] = useState<SongObject[]>([]);
-  const [refreshQueue, setRefreshQueue] = useState<boolean>(true);
-  const [loading, setLoading] = useState(false);
-  const pathname = usePathname();
+  const { queueData, loading, refreshQueue } = useQueue();
 
-  usePusher(() => setRefreshQueue(true));
-
-  useEffect(() => {
-    if (!pathname || !refreshQueue) return;
-    setLoading(true);
-    const userProfile = getUserProfileFromLocal();
-    const hostname = pathname.substring(1).split('/')[0];
-    getQueue(hostname, userProfile).then(res => {
-      if (res) setQueueData(res);
-      setLoading(false);
-    });
-    setRefreshQueue(false);
-  }, [pathname, refreshQueue]);
+  usePusher(() => refreshQueue());
 
   const EmptyQueue = () => {
     return (
@@ -124,7 +40,6 @@ export default function Queue() {
       ) : queueData.length === 0 ? (
         <EmptyQueue />
       ) : (
-        // <div className="pb-[60px] text-white relative z-50 flex flex-col items-center min-h-screen font-thin">
         <div className="pb-[60px] max-w-screen-sm m-auto pb-36 text-white relative z-50 flex flex-col items-center min-h-screen font-thin">
           {queueData.map((song, key) => (
             <QueueSong key={key} song={song} />
